requirements/common.txt
Django>=1.8,<1.9
mysql-python
django-admin-bootstrapped<=2.5

#partial dates
git+git://github.com/dracos/django-date-extensions.git
django-ajax-selects
django-multiselectfield

python-igraph
lxml

-e git://github.com/emory-libraries/django-eultheme#egg=django_eultheme
django-widget-tweaks
django-stdimage

